fix(build): make mixed float-double conversions explicit
Add explicit casts and drop the float suffix on constants across the
calibration and magneto paths so implicit float/double conversions no
longer trigger compiler warnings and the intended precision of each
operation is unambiguous.

+++ src/sensor/magneto/magneto1_4.c
@@ -347,13 +347,13 @@ int magneto_current_calibration(float BAinv[4][3], double *ata, double norm_sum,
     // Check the entire candidate before the first caller-visible write. A
     // finite double can still overflow or lose its positive diagonal in float.
     for (int i = 0; i < 3; i++) {
-        if (!isfinite(B[i]) || fabs(B[i]) > FLT_MAX) {
+        if (!isfinite(B[i]) || fabs(B[i]) > (double)FLT_MAX) {
             err = -ERANGE;
             goto cleanup;
         }
     }
     for (int i = 0; i < 9; i++) {
-        if (!isfinite(A_1[i]) || fabs(A_1[i]) > FLT_MAX) {
+        if (!isfinite(A_1[i]) || fabs(A_1[i]) > (double)FLT_MAX) {
             err = -ERANGE;
             goto cleanup;
         }
